fix(template): address regression in string functions for template engine
It appears that `Sprig` flips the inputs of the `contains`, `hasPrefix`, `replace`
and `hasSuffix` `strings` methods. This appears to be the cause of the regression.

See: https://github.com/Masterminds/sprig/blob/master/functions.go#L149-L152

This results in a regression against the previous implementations of the template
functions.

Signed-off-by: Derek Smith <derek@clokwork.net>